Robolectric + Instrumentation の戦いに敗れたのでDBのテストは素直にandroidTestに移動
