+++ app/build.gradle
@@ -35,6 +35,7 @@ android {
     }
     sourceSets {
         test.java.srcDirs += 'src/test/kotlin'
+        androidTest.assets.srcDirs += files("$projectDir/schemas".toString())
     }
     dataBinding {
         enabled true
@@ -105,8 +106,8 @@ dependencies {
     testImplementation 'org.hamcrest:hamcrest-library:1.3'
     testImplementation "org.jetbrains.kotlin:kotlin-test:$kotlin_version"
     testImplementation "org.jetbrains.kotlin:kotlin-test-junit:$kotlin_version"
-    testImplementation "android.arch.persistence.room:testing:${arch_lib_version}"
-    testImplementation 'com.github.gfx.android.robolectricinstrumentation:robolectric-instrumentation:3.1.4'
+
+    androidTestImplementation "android.arch.persistence.room:testing:${arch_lib_version}"
 }
 
 apply plugin: 'com.google.gms.google-services'
